Add files via upload
These python codes automate the process of soil loss estimation from the modified MMF model. The whole process and steps to calculate the soil loss and sedimentation yield from a given watershed are published in the following paper:
"Modified MMF (Morgan–Morgan–Finney) model for evaluating effects of crops and vegetation cover on soil erosion"
https://onlinelibrary.wiley.com/doi/abs/10.1002/esp.1530 

Please carefully read the paper and the steps to use these codes for soil loss and sedimentation yield estimation for the area of interest. Users may also need to modify these codes based on their area of study.